Guard mtl-related instances for ErrorT/ListT with appropriate CPP
Previously, all instances involving `ErrorT` and `ListT` were guarded with
`!MIN_VERSION_transformers(0,6,0)`. This turns out not to be enough, however,
as the instances involving classes from `mtl` also need to be guarded with
`!MIN_VERSION_mtl(2,3,0)`. This is because while `mtl-2.3.*` does build with
`transformers-0.5.*`, it does not provide instances for `ErrorT` or `ListT`,
which could result in the build failures observed in #115.

Fixes #115.