Implement Data.Constraint.{Nat,Symbol}'s magic using withKnown{Nat,Symbol} when possible
This reduces the amount of code that has to be `unsafeCoerce`d down to just the
`unsafeSNat` and `unsafeSSymbol` functions, which I have added to
`Data.Constraint.Unsafe`. (If https://gitlab.haskell.org/ghc/ghc/-/issues/23478
were implemented, then we could do away with even those uses of
`unsafeCoerce`.)

Fixes #114.